feat(tarefa): creating the TarefaRepository class

+++ src/main/java/com/eranitor/tcc/entity/Tarefa.java
@@ -16,7 +16,7 @@ public class Tarefa {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
     @Column (name = "idtarefa")
-    private Integer idTarefa;
+    private Long idTarefa;
 
     @ManyToOne
     @JoinColumn (name = "materia_idmateria")
